Replace continuous_mlp_q_function with model
